requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
asynctest==0.13.0
atpublic==2.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.4
cassandra-driver==3.25.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==35.0.0
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.6.0
dpath==2.0.5
dulwich==0.20.26
dvc==2.8.3
flatten-dict==0.4.2
flufl.lock==6.0
frozenlist==1.2.0
fsspec==2021.11.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
geomet==0.2.1.post1
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
grandalf==0.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.4.0
jaraco.structures==2.1.0
jaraco.text==3.6.0
jaraco.ui==2.3.0
jaraco.windows==5.7.0
joblib==1.1.0
mailchecker==4.0.16
more-itertools==8.11.0
multidict==5.2.0
nanotime==0.5.2
networkx==2.6.3
numpy==1.21.4
oauth2client==4.1.3
packaging==21.3
pandas==1.3.4
path==16.2.0
pathspec==0.9.0
phonenumbers==8.12.37
ply==3.11
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyDrive2==1.10.0
pygit2==1.7.0
Pygments==2.10.0
pygtrie==2.4.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rich==10.14.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.2
shortuuid==1.0.8
shtab==1.5.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
voluptuous==0.12.2
wcwidth==0.2.5
wincertstore==0.2
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
